refactor(create page): move business logic into hook
